Reorg and add select tests
